+++ CHANGELOG.md
@@ -6,11 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## ESLintPresets [Unreleased]
 ### Fixed
-  - Vulnerabilities in dependencies
-  - Broken ruleset by replacing `eslint-plugin-deprecation` with `@delagen/eslint-plugin-deprecation`
+  - All vulnerabilities in dependencies
 
 ### Updated
+  - Linting environment
   - All dependencies
+  - Deprecated rules
+    - `@typescript-eslint/no-parameter-properties` has been replaced by `@typescript-eslint/parameter-properties`
+
+### Added
+  - Reports for the use of deprecated rules while testing
 
 [Show differences](https://github.com/manuth/ESLintPresets/compare/v3.2.6...dev)