+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## ESLintPresets 3.2.8
 ### Fixed
   - All vulnerabilities in dependencies
+  - Test timeouts
 
 ### Updated
   - Linting environment